refactor: extract common gen script helpers into gen_common.py
Move _ensure_inductor_compiler, _load_custom_ops, and print_cpp_values
into gen_common.py. Use CXX env var instead of hardcoded compiler
fallback list. All gen scripts now import from gen_common.
