+++ source/tests/infer/gen_dpa3.py
@@ -8,9 +8,7 @@ Also prints reference values for C++ tests (PBC and NoPbc).
 """
 
 import copy
-import glob
 import os
-import shutil
 import sys
 
 import numpy as np
@@ -18,86 +16,11 @@ import numpy as np
 # Ensure the source tree is on the path
 sys.path.insert(0, os.path.join(os.path.dirname(__file__), "..", "..", ".."))
 
-
-def _ensure_inductor_compiler():
-    """Ensure torch._inductor can find a C++ compiler.
-
-    torch._inductor searches for 'g++' by default.  On some CI images only
-    versioned binaries (e.g. g++-11) or 'c++' exist.  Fall back to those.
-    """
-    import torch._inductor.config as inductor_config
-
-    search = inductor_config.cpp.cxx
-    if isinstance(search, (list, tuple)):
-        search = list(search)
-    else:
-        search = [search]
-    # Append common fallbacks that are not in the default search list
-    for fallback in ["c++", "g++-14", "g++-13", "g++-12", "g++-11"]:
-        if fallback not in search and shutil.which(fallback):
-            search.append(fallback)
-    inductor_config.cpp.cxx = tuple(search)
-    # Clear LD_PRELOAD so compiler subprocesses don't inherit the LSAN runtime,
-    # which causes false leak reports and non-zero exit codes in g++.
-    os.environ.pop("LD_PRELOAD", None)
-
-
-def _load_custom_ops():
-    """Load custom op library if not already registered.
-
-    Must be called AFTER importing deepmd (which may register ops from the
-    pip-installed library) to avoid double-registration crashes.
-    """
-    import torch
-
-    if hasattr(torch.ops, "deepmd") and hasattr(torch.ops.deepmd, "border_op"):
-        return
-    _search_base = os.path.realpath(os.path.dirname(__file__))
-    for pattern in [
-        os.path.join(
-            _search_base, "..", "..", "..", "build*", "op", "pt", "libdeepmd_op_pt.so"
-        ),
-        os.path.join(
-            _search_base, "..", "..", "build*", "op", "pt", "libdeepmd_op_pt.so"
-        ),
-    ]:
-        libs = glob.glob(pattern)
-        if libs:
-            try:
-                torch.ops.load_library(libs[0])
-            except (OSError, RuntimeError) as e:
-                import warnings
-
-                warnings.warn(
-                    f"Custom op library not loaded: {e}",
-                    stacklevel=2,
-                )
-            break
-
-
-def print_cpp_values(label, ae, f, av):
-    """Print C++ reference arrays."""
-    print(f"\n// ---- {label} ----")  # noqa: T201
-    atom_energy = ae[0, :, 0]
-    print("  std::vector<VALUETYPE> expected_e = {")  # noqa: T201
-    for ii, e in enumerate(atom_energy):
-        comma = "," if ii < len(atom_energy) - 1 else ""
-        print(f"      {e:.18e}{comma}")  # noqa: T201
-    print("  };")  # noqa: T201
-
-    print("  std::vector<VALUETYPE> expected_f = {")  # noqa: T201
-    force_flat = f[0].flatten()
-    for ii, fv in enumerate(force_flat):
-        comma = "," if ii < len(force_flat) - 1 else ""
-        print(f"      {fv:.18e}{comma}")  # noqa: T201
-    print("  };")  # noqa: T201
-
-    print("  std::vector<VALUETYPE> expected_v = {")  # noqa: T201
-    virial_flat = av[0].flatten()
-    for ii, v in enumerate(virial_flat):
-        comma = "," if ii < len(virial_flat) - 1 else ""
-        print(f"      {v:.18e}{comma}")  # noqa: T201
-    print("  };")  # noqa: T201
+from gen_common import (
+    ensure_inductor_compiler,
+    load_custom_ops,
+    print_cpp_values,
+)
 
 
 def main():
@@ -105,7 +28,7 @@ def main():
         get_model,
     )
 
-    _ensure_inductor_compiler()
+    ensure_inductor_compiler()
 
     # ---- 1. DPA3 model config (small, fast to compile) ----
     config = {
@@ -157,7 +80,7 @@ def main():
     )
 
     # Load custom ops after deepmd.pt import to avoid double registration
-    _load_custom_ops()
+    load_custom_ops()
 
     base_dir = os.path.dirname(__file__)
