+++ source/tests/infer/gen_sea.py
@@ -6,64 +6,16 @@ Converts the existing deeppot_sea.pth (checked into git) to .pt2 format
 via serialize -> deserialize. Reference values are already in the C++ test files.
 """
 
-import glob
 import os
-import shutil
 import sys
 
 # Ensure the source tree is on the path
 sys.path.insert(0, os.path.join(os.path.dirname(__file__), "..", "..", ".."))
 
-
-def _ensure_inductor_compiler():
-    """Ensure torch._inductor can find a C++ compiler.
-
-    torch._inductor searches for 'g++' by default.  On some CI images only
-    versioned binaries (e.g. g++-11) or 'c++' exist.  Fall back to those.
-    """
-    import torch._inductor.config as inductor_config
-
-    search = inductor_config.cpp.cxx
-    if isinstance(search, (list, tuple)):
-        search = list(search)
-    else:
-        search = [search]
-    # Append common fallbacks that are not in the default search list
-    for fallback in ["c++", "g++-14", "g++-13", "g++-12", "g++-11"]:
-        if fallback not in search and shutil.which(fallback):
-            search.append(fallback)
-    inductor_config.cpp.cxx = tuple(search)
-    # Clear LD_PRELOAD so compiler subprocesses don't inherit the LSAN runtime,
-    # which causes false leak reports and non-zero exit codes in g++.
-    os.environ.pop("LD_PRELOAD", None)
-
-
-def _load_custom_ops():
-    """Load custom op library if not already registered.
-
-    Must be called AFTER importing deepmd (which may register ops from the
-    pip-installed library) to avoid double-registration crashes.
-    """
-    import torch
-
-    if hasattr(torch.ops, "deepmd") and hasattr(torch.ops.deepmd, "border_op"):
-        return
-    _search_base = os.path.realpath(os.path.dirname(__file__))
-    for pattern in [
-        os.path.join(
-            _search_base, "..", "..", "..", "build*", "op", "pt", "libdeepmd_op_pt.so"
-        ),
-        os.path.join(
-            _search_base, "..", "..", "build*", "op", "pt", "libdeepmd_op_pt.so"
-        ),
-    ]:
-        libs = glob.glob(pattern)
-        if libs:
-            try:
-                torch.ops.load_library(libs[0])
-            except Exception as e:
-                print(f"NOTE: custom op library not loaded ({e})", file=sys.stderr)  # noqa: T201
-            break
+from gen_common import (
+    ensure_inductor_compiler,
+    load_custom_ops,
+)
 
 
 def main():
@@ -79,8 +31,8 @@ def main():
     )
 
     # Load custom ops after deepmd.pt import to avoid double registration
-    _load_custom_ops()
-    _ensure_inductor_compiler()
+    load_custom_ops()
+    ensure_inductor_compiler()
 
     base_dir = os.path.dirname(__file__)
     pth_path = os.path.join(base_dir, "deeppot_sea.pth")
